Added Cmds to Help + Clear requires Perms

+++ commands/clear.js
@@ -10,12 +10,14 @@ module.exports = {
       return msg.reply("that doesn't seem to be a valid number.");
     } else if (amount <= 1 || amount > 100) {
       return msg.reply("you need to input a number between 1 and 99.");
-    } 
-    msg.channel.bulkDelete(amount, true).catch((err) => {
-      console.error(err);
-      msg.channel.send(
-        "there was an error trying to delete messages in this channel!"
-      );
-    });
+    }
+    if (msg.author.hasPermission("ADMINISTRATOR")) {
+      msg.channel.bulkDelete(amount, true).catch((err) => {
+        console.error(err);
+        msg.channel.send(
+          "there was an error trying to delete messages in this channel!"
+        );
+      });
+    }
   },
 };
